+++ frontend/src/components/bussiness/ms-breadcrumb/types.ts
@@ -1,0 +1,6 @@
+import type { RouteRecordName } from 'vue-router';
+
+export interface BreadcrumbItem {
+  name: RouteRecordName;
+  locale: string;
+}
