+++ frontend/src/locale/en-US/index.ts
@@ -1,33 +1,34 @@
 import dayjsLocale from 'dayjs/locale/en';
 import localeSettings from './settings';
 import sys from './sys';
-import localeMessageBox from '@/components/pure/message-box/locale/en-US';
-import minder from '@/components/pure/minder-editor/locale/en-US';
-import localeLogin from '@/views/login/locale/en-US';
-import localeTable from '@/components/pure/ms-table/locale/en-US';
-import localeApiTest from '@/views/api-test/locale/en-US';
-import localeSystem from '@/views/system/locale/en-US';
-import baseLocale from '@/views/base/locale/en-US';
 
+const _Cmodules: any = import.meta.glob('../../components/**/locale/en-US.ts', { eager: true });
+const _Vmodules: any = import.meta.glob('../../views/**/locale/en-US.ts', { eager: true });
+let result = {};
+Object.keys(_Cmodules).forEach((key) => {
+  const defaultModule = _Cmodules[key as any].default;
+  if (!defaultModule) return;
+  result = { ...result, ...defaultModule };
+});
+Object.keys(_Vmodules).forEach((key) => {
+  const defaultModule = _Vmodules[key as any].default;
+  if (!defaultModule) return;
+  result = { ...result, ...defaultModule };
+});
 export default {
   message: {
     'menu.apiTest': 'Api Test',
     'menu.settings': 'System Settings',
     'menu.settings.system': 'System',
     'menu.settings.organization': 'Organization',
-    'menu.settings.usergroup': 'User Group',
-    'menu.settings.user': 'User',
-    'menu.settings.organizationAndProject': 'Org & Project',
+    'menu.settings.system.usergroup': 'User Group',
+    'menu.settings.system.user': 'User',
+    'menu.settings.system.organizationAndProject': 'Org & Project',
+    'menu.settings.system.resourcePool': 'Resource Pool',
     'navbar.action.locale': 'Switch to English',
     ...sys,
     ...localeSettings,
-    ...localeMessageBox,
-    ...localeLogin,
-    ...minder,
-    ...localeTable,
-    ...localeApiTest,
-    ...localeSystem,
-    ...baseLocale,
+    ...result,
   },
   dayjsLocale,
   dayjsLocaleName: 'en-US',
