--- frontend/src/locale/en-US/system.ts
@@ -1,1 +1,0 @@
-export default {};
