--- frontend/src/locale/zh-CN/system.ts
@@ -1,1 +1,0 @@
-export default {};
