+++ frontend/src/views/base/locale/en-US.ts
