+++ frontend/src/views/base/locale/zh-CN.ts
