+++ frontend/src/views/system/user/locale/en-US.ts
@@ -1,47 +1,4 @@
 export default {
-  'system': {
-    userGroup: {
-      searchHolder: 'Please input user group name',
-      inSystem: 'In system',
-      customUserGroup: 'Custom user group',
-      addSystemUserGroup: 'Add system user group',
-      addSystemUser: 'Add system user',
-      addUser: 'Add user',
-      emptyUserGroup:
-        'There are currently no custom user groups available. Please click "Create" or "+" above to create a user group',
-      rename: 'Rename',
-      changeAuthScope: 'Change auth scope',
-      delete: 'Delete',
-      user: 'user',
-      add: 'Add',
-      cancel: 'Cancel',
-      create: 'Create',
-      createUserGroup: 'create user group',
-      userGroupName: 'User group name',
-      authScope: 'Auth scope',
-      auth: 'Auth',
-      deleteSuccess: 'Delete success',
-      revokeSuccess: 'Revoke success',
-      revoke: 'revoke',
-      pleaseSelectUser: 'Please select user',
-      pleaseInputUserGroupName: 'Please input user group name,and not equal to system user group name',
-      userGroupNameIsExist: `Usergroup {name} is exist`,
-      pleaseSelectAuthScope: 'Please select auth scope',
-      userGroupNameIsNotNone: 'User group name is not none',
-      authScopeIsNotNone: 'Auth scope is not none',
-      userIsNotNone: 'User is not none',
-      confirm: 'Confirm',
-      global: 'Global',
-      searchPlacehoder: 'Search by ID/Name',
-      SYSTEM: '系统',
-      PROJECT: '项目',
-      ORGANIZATION: '组织',
-      isDeleteUserGroup: 'Delete or not: {name}?',
-      beforeDeleteUserGroup:
-        'After deletion, the project data under the organization will be deleted together. Please operate with caution!',
-      confirmDelete: 'Confirm delete',
-    },
-  },
   'system.user.createUser': 'Create User',
   'system.user.emailInvite': 'Email Invite',
   'system.user.importUser': 'Import User',
