+++ frontend/src/views/system/usergroup/locale/zh-CN.ts
@@ -1,0 +1,44 @@
+export default {
+  system: {
+    userGroup: {
+      global: '全局用户组',
+      searchHolder: '请输入用户组名称',
+      inSystem: '系统内置',
+      customUserGroup: '自定义用户组',
+      addSystemUserGroup: '添加系统用户组',
+      addSystemUser: '添加系统用户',
+      addUser: '添加成员',
+      emptyUserGroup: '暂无自定义用户组，请点击上方创建或 “+” 创建用户组',
+      rename: '重命名',
+      changeAuthScope: '修改权限范围',
+      delete: '删除',
+      user: '成员',
+      add: '添加',
+      confirm: '确认',
+      create: '创建',
+      cancel: '取消',
+      createUserGroup: '创建用户组',
+      userGroupName: '用户组名称',
+      authScope: '权限范围',
+      auth: '权限',
+      deleteSuccess: '删除成功',
+      revokeSuccess: '撤销成功',
+      revoke: '撤销',
+      pleaseSelectUserGroup: '请选择用户组',
+      pleaseSelectUser: '请选择成员',
+      userGroupNameIsNotNone: '用户组名称不能为空',
+      authScopeIsNotNone: '权限范围不能为空',
+      userIsNotNone: '成员不能为空',
+      pleaseInputUserGroupName: '请输入用户组名称，且不与其他用户组名称重复',
+      userGroupNameIsExist: `用户组{name}已存在`,
+      pleaseSelectAuthScope: '请选择用户组所属的权限范围',
+      searchPlacehoder: '通过ID/名称搜索',
+      SYSTEM: '系统',
+      PROJECT: '项目',
+      ORGANIZATION: '组织',
+      isDeleteUserGroup: '是否删除: {name}?',
+      beforeDeleteUserGroup: '删除后，该组织下的项目数据将一起删除，请谨慎操作！',
+      confirmDelete: '确认删除',
+    },
+  },
+};
